- Split onChatMessage into own function - ChatStatusOverlay has own file and class - Old unused code deleted - Most bloated comments removed
